streamlit/requirements.txt
streamlit
streamlit-autorefresh
psycopg2-binary
pandas
plotly
python-dotenv
SQLAlchemy